refactor: rename internal/gitindex package to internal/indexer
