+++ internal/indexer/indexer_test.go
@@ -1,4 +1,4 @@
-package gitindex
+package indexer
 
 import (
 	"archive/zip"
